Send json responses back to frontend with fully qualified state
The backend now sends a json response back to the frontend with a fully
qualified state so the frontend can update based on that.

For example, when adding a new todo, only the text that was gathered
from the frontend is sent to the backend and the backend now generates
the actual todo with all of its properties set and sends that back to
the frontend to be displayed in the list.