Restructure the backend to add testing
The backend code now has some testing for the db.py and app.py files.
The backend's structure had to be moved around a little bit in order for
this to happen.

All of the server files now live in a `server` module (server folder)
and the tests live in the `tests` module.

In order to run the tests run either run the `pipenv shell` or just
prepend `pipenv run` to `pytest --cov=server tests\`.

Unfortunately, at this time there are some warnings based on the
packages that come with Werkzug, but they can be safely ignored for the
time being.